type_c/rag_api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
pgvector==0.2.5
# tiny helper for prompt templates
langchain-core==0.2.0
python-dotenv==1.0.1
spacy==3.8
tiktoken
groq
psycopg2-binary
neo4j
redis